move encode, use format_args
