+++ src/sync.rs
@@ -44,60 +44,3 @@ pub fn connect_easy(
     let config = simple_user_config(name, token).unwrap();
     connect(&config, secure)
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::encode::*;
-
-    #[test]
-    fn encode_raw() {
-        let mut out = vec![];
-        super::encode(&raw("PRIVMSG #test :hello world"), &mut out).unwrap();
-        assert_eq!(out, b"PRIVMSG #test :hello world\r\n");
-    }
-
-    #[test]
-    fn encode_ping() {
-        let mut out = vec![];
-        super::encode(&ping("123456789"), &mut out).unwrap();
-        assert_eq!(out, b"PING 123456789\r\n");
-    }
-
-    #[test]
-    fn encode_pong() {
-        let mut out = vec![];
-        super::encode(&pong("123456789"), &mut out).unwrap();
-        assert_eq!(out, b"PONG :123456789\r\n");
-    }
-
-    #[test]
-    fn encode_join() {
-        let mut out = vec![];
-        super::encode(&join("#museun"), &mut out).unwrap();
-        assert_eq!(out, b"JOIN #museun\r\n");
-    }
-
-    #[test]
-    fn encode_part() {
-        let mut out = vec![];
-        super::encode(&part("#museun"), &mut out).unwrap();
-        assert_eq!(out, b"PART #museun\r\n");
-    }
-
-    #[test]
-    fn encode_privmsg() {
-        let mut out = vec![];
-        super::encode(&privmsg("#museun", "this is a test of a line"), &mut out).unwrap();
-        assert_eq!(
-            out,
-            "PRIVMSG #museun :this is a test of a line\r\n".as_bytes()
-        );
-
-        let mut out = vec![];
-        super::encode(&privmsg("#museun", &"foo ".repeat(500)), &mut out).unwrap();
-        assert_eq!(
-            out,
-            format!("PRIVMSG #museun :{}\r\n", &"foo ".repeat(500)).as_bytes()
-        );
-    }
-}
